hold strong refs to app in command objects
This fixes the changes in Minion 9.07, which weakened various object
attributes to try to reduce memory leaks. Indeed, we had to weaken the
app reference ourselves to prevent a memory leak, so Minion is now doing
this for us.

This fixes the `Reference is already weak` warning and the test failure
`Can't call method "minion" on an undefined value` that affected the job
and worker command tests.

Thanks @eserte for reporting this issue!

Fixes #4